Fixes: default plugin name for msvc
On windows the file ending is .dll and for msvc there is no "lib" added
to the front.

The wrong name caused a bug on windows because the default plugin name is used
in a default constructor and then compared to the loaded plugin causing the
auto matching of system language to fail. The fallback is that the first
language in the list (arabic) is selected.